Reorg for convention. Update refs and build script.
