+++ IFC-dotnet.sln
@@ -1,0 +1,48 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IFC-dotnet", "src\IFC-dotnet\IFC-dotnet.csproj", "{301045AC-77F1-4B10-B668-0213CB53A2AB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IFC-dotnet-test", "tests\IFC-dotnet-test\IFC-dotnet-test.csproj", "{3FA9536F-5D93-49B0-88E6-365400E8BB60}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|x64.ActiveCfg = Debug|x64
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|x64.Build.0 = Debug|x64
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|x86.ActiveCfg = Debug|x86
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Debug|x86.Build.0 = Debug|x86
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|x64.ActiveCfg = Release|x64
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|x64.Build.0 = Release|x64
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|x86.ActiveCfg = Release|x86
+		{301045AC-77F1-4B10-B668-0213CB53A2AB}.Release|x86.Build.0 = Release|x86
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|x64.ActiveCfg = Debug|x64
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|x64.Build.0 = Debug|x64
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|x86.ActiveCfg = Debug|x86
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Debug|x86.Build.0 = Debug|x86
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|x64.ActiveCfg = Release|x64
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|x64.Build.0 = Release|x64
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|x86.ActiveCfg = Release|x86
+		{3FA9536F-5D93-49B0-88E6-365400E8BB60}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+EndGlobal
