--- src/build.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-cd IFC-dotnet
-antlr4 -Dlanguage=CSharp -package STEP -o ./antlr STEP.g4
-cd ..
-dotnet build
-cd IFC-dotnet-test
-dotnet xunit
-cd ..
